SDK-3306: Error class hierarchy (#186)
This introduces the core classes for this library, namely:
- `InruptClientError`, the superclass for all Inrupt client related errors,
-  `ClientHttpError`, the base class for HTTP errors

Note that classes for specific HTTP errors (Unauthenticated, Not Found...) are not included yet, and will be added later.

`ClientHttpError` extends the base `InruptClientError` by implementing two interfaces, `WithProblemDetails` and `WithErrorResponse`, respectively associated to their type guards `hasProblemDetails` and `hasErrorResponse`. Both the `.problemDetails` and `.response` accessors are marked as read-only, and immutability is enforced at runtime for libraries not using static analysis based on TS.

The constructor of `ClientHttpError` takes in a response body and a subset of the native `Response` metadata, instead of taking a simple `Response` instance, because reading the `Response` body is an asynchronous operation, which isn't possible when building an object. It is expected that the parsing will be done by the caller in an async context, and the result will be passed to the `ClientHttpError` constructor, which is not the initial design.

For the time being, I'm not doing any submodule exports (e.g. `@inrupt/solid-client-error/http`), but this is something we might eventually consider for tree-shaking.

Co-authored-by: Pete Edwards <edwardsph@users.noreply.github.com>